Update new 22/9/2019

+++ ElectricStore/Logic/LogicExtension/ProductDetailRepository.cs
@@ -9,11 +9,16 @@ namespace ElectricStore.Logic.LogicExtension
     public class ProductDetailRepository : IProductDetailRepository
     {
         private LaptopStoreEntities context = new LaptopStoreEntities();
-        public void Delete(int id)
+        public string Delete(int id)
         {
             ProductDetail product_detail = context.ProductDetails.Find(id);
+            if(product_detail is null)
+            {
+                return "Xóa chi tiết sản phẩm không thành công";
+            }
             context.ProductDetails.Remove(product_detail);
             context.SaveChanges();
+            return "Xóa chi tiết sản phẩm thành công";
         }
 
         public ProductDetail GetById(int id)
